MeshLodGenerator: move compression into OutputProvider base class
