+++ Components/MeshLodGenerator/include/OgreLodOutputProvider.h
@@ -38,16 +38,37 @@ namespace Ogre
 
 class _OgreLodExport LodOutputProvider {
 public:
+    LodOutputProvider(bool useCompression = false);
     virtual ~LodOutputProvider() {}
-    virtual void prepare(LodData* data) = 0;
-    virtual void finalize(LodData* data) = 0;
+    virtual void prepare(LodData* data);
+    virtual void finalize(LodData* data);
     virtual void bakeManualLodLevel(LodData* data, String& manualMeshName, int lodIndex);
     virtual void bakeLodLevel(LodData* data, int lodIndex);
-    virtual void triangleRemoved(LodData* data, LodData::Triangle* tri){}
-    virtual void triangleChanged(LodData* data, LodData::Triangle* tri){}
+    virtual void triangleRemoved(LodData* data, LodData::Triangle* tri);
+    virtual void triangleChanged(LodData* data, LodData::Triangle* tri);
     virtual void inject(){}
 
 protected:
+    bool mUseCompression;
+    
+    struct TriangleCache {
+        unsigned int vertexID[3];
+        bool vertexChanged;
+    };
+
+    typedef std::vector<TriangleCache> TriangleCacheList;
+
+    /// First pass will create the mTriangleCacheList and second pass will use it.
+    /// This is required, because the triangles from first pass will be changed and we need to keep the information.
+    TriangleCacheList mTriangleCacheList;
+
+    bool mFirstBufferPass;
+    int mLastIndexBufferID;
+
+    void bakeUncompressed(LodData* data, int lodIndex);
+    void bakeFirstPass(LodData* data, int lodIndex);
+    void bakeSecondPass(LodData* data, int lodIndex);
+
     
     // TODO: remove implementation and make pure virtual. These are just to make the compressed version work.
     virtual size_t getSubMeshCount() { return 0; } // = 0;
